refactor(rules): use more explicit naming
